requirements.txt
python-dotenv
fastapi
psycopg2-binary
musicbrainzngs
uvicorn
sqlalchemy
psycopg2-binary
pandas
